% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bi_all_data.R
\name{bi_all_data}
\alias{bi_all_data}
\title{Retrieve both image data and metadata}
\usage{
bi_all_data(
  identifier = NULL,
  region = c(0L, 0L, 500L, 500L),
  size = "full",
  rotation = 0,
  quality = c("native", "color", "gray", "bitonal"),
  format = c("jpg", "tif", "png", "gif", "jp2", "pdf")
)
}
\arguments{
\item{identifier}{[\verb{character(1+)}]\cr{}
one or multiple identifier(s) of the document that begins
with \code{"ark:/"}}

\item{region}{[\code{character(1)} or \code{integer(4)}]\cr{}
either \code{"full"} to retrieve the full image or a vector of
four integers that defines the region to extract from the
image. The vector has the form \verb{(x0, y0, w, h)} with
\verb{x0, y0} the coordinates of the top left point of
the extracted region (with the origin at the top left of
the image), \code{w} the width of the extracted region and
\code{h} the height of the extracted region.}

\item{size}{[\code{character(1)} or \code{integer(2)}]\cr{}
either \code{"full"} to retrieve the full region or a vector of
two integers that defines the size of the final image}

\item{rotation}{[\code{numeric(1)}]\cr{}
an integer that defines the rotation in degrees to apply
to the final image (\code{0} for no rotation)}

\item{quality}{[\code{character(1)}]\cr{}
\itemize{
\item \code{"native"} for native colors of the image or,
\item \code{"color"} for the image in color,
\item \code{"gray"} for levels of gray,
\item \code{"bitonal"} for only black and white pixels.
}}

\item{format}{[\code{character(1)}]\cr{}
\itemize{
\item \code{"jpg"} for JPEG image,
\item \code{"tif"} for TIFF image,
\item \code{"png"} for PNG image,
\item \code{"gif"} for GIF image,
\item \code{"jp2"} for JP2 image,
\item \code{"pdf"} for PDF image.
}}
}
\value{
a tibble with both the image and the metadata, with one row per
identifier
}
\description{
This function queries both data and associated metadata from a BNF image
using specified parameters and return the information in a tidy data.frame
}
