% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bi_image.R
\name{bi_image}
\alias{bi_image}
\title{Retrieve image from BNF}
\usage{
bi_image(identifier = NULL, region = c(0L, 0L, 500L, 500L),
  size = "full", rotation = 0, quality = c("native", "bitonal"),
  format = c("jpg", "png"))
}
\arguments{
\item{identifier}{[\code{character(1)}]\cr{}
identifier of the document that begins with \code{"ark:/"}}

\item{region}{[\code{character(1)} or \code{integer(4)}]\cr{}
either \code{"full"} to retrieve the full image or a vector of
four integers that defines the region to extract from the
image}

\item{size}{[\code{character(1)} or \code{integer(2)}]\cr{}
either \code{"full"} to retrieve the full region or a vector of
two integers that defines the size of the final image}

\item{rotation}{[\code{numeric(1)}]\cr{}
an integer that defines the rotation in degrees to apply
to the final image (\code{0} for no rotation)}

\item{quality}{[\code{character(1)}]\cr{}
\code{"native"} for native colors of the image or
\code{"bitonal"} for levels of gray}

\item{format}{[\code{character(1)}]\cr{}
\code{"jpg"} for JPEG image
\code{"png"} for PNG image}
}
\value{
a matrix of the image
}
\description{
Retrieve image from BNF
}
