% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bi_metadata.R
\name{bi_metadata}
\alias{bi_metadata}
\title{Retrieve BnF image metadata}
\usage{
bi_metadata(identifier = NULL)
}
\arguments{
\item{identifier}{[\code{character(1)}]\cr{}
identifier of the document that begins with \code{"ark:/"}}
}
\description{
Retrieve BnF image metadata
}
\details{
By default \code{bnfimage} implements rate limitation through the
\href{https://cran.r-project.org/package=ratelimitr}{\code{ratelimitr} package}.
The query are limited to a maximum of 1 every 3 seconds to avoid being
blocked by the BnF server. This is the limit at which the BnF server
considers the queries to be malicious. For the moment this limitation
cannot be lifted.
}
