% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bi_image.R
\name{bi_image}
\alias{bi_image}
\title{Retrieve image from BNF}
\usage{
bi_image(
  identifier = NULL,
  region = c(0L, 0L, 500L, 500L),
  size = "full",
  rotation = 0,
  quality = c("native", "color", "gray", "bitonal"),
  format = c("jpg", "tif", "png", "gif", "jp2", "pdf")
)
}
\arguments{
\item{identifier}{[\code{character(1)}]\cr{}
identifier of the document that begins with \code{"ark:/"}}

\item{region}{[\code{character(1)} or \code{integer(4)}]\cr{}
either \code{"full"} to retrieve the full image or a vector of
four integers that defines the region to extract from the
image. The vector has the form \verb{(x0, y0, w, h)} with
\verb{x0, y0} the coordinates of the top left point of
the extracted region (with the origin at the top left of
the image), \code{w} the width of the extracted region and
\code{h} the height of the extracted region.}

\item{size}{[\code{character(1)} or \code{integer(2)}]\cr{}
either \code{"full"} to retrieve the full region or a vector of
two integers that defines the size of the final image}

\item{rotation}{[\code{numeric(1)}]\cr{}
an integer that defines the rotation in degrees to apply
to the final image (\code{0} for no rotation)}

\item{quality}{[\code{character(1)}]\cr{}
\itemize{
\item \code{"native"} for native colors of the image or,
\item \code{"color"} for the image in color,
\item \code{"gray"} for levels of gray,
\item \code{"bitonal"} for only black and white pixels.
}}

\item{format}{[\code{character(1)}]\cr{}
\itemize{
\item \code{"jpg"} for JPEG image,
\item \code{"tif"} for TIFF image,
\item \code{"png"} for PNG image,
\item \code{"gif"} for GIF image,
\item \code{"jp2"} for JP2 image,
\item \code{"pdf"} for PDF image.
}}
}
\value{
a matrix of the image
}
\description{
Retrieve image from BNF
}
\details{
By default \code{bnfimage} implements rate limitation through the
\href{https://cran.r-project.org/package=ratelimitr}{\code{ratelimitr} package}.
The query are limited to a maximum of 1 every 3 seconds to avoid being
blocked by the BnF server. This is the limit at which the BnF server
considers the queries to be malicious. For the moment this limitation
cannot be lifted.
}
