% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bi_metadata.R
\name{bi_metadata}
\alias{bi_metadata}
\title{Retrieve BnF image metadata}
\usage{
bi_metadata(identifier = NULL)
}
\arguments{
\item{identifier}{[\code{character(1)}]\cr{}
identifier of the document that begins with \code{"ark:/"}}
}
\description{
Retrieve BnF image metadata
}
